part: switch to json loading
Every time we do a 'make install' we also update the mime database (because we
install mimetypes) and this causes the ksycoca file to be regenerated the
next time we run Ark (because we load the Part using KService).

This is annoying and clutters the console output with ksycoca messages. We can use instead the new
json loading mechanism like we already do with our plugins.

Regarding ark_part.desktop, we continue to install it because Konqueror is
probably using it.

Differential Revision: D2140