Implement a custom AddDialog
This diff implements an AddDialog class that is used instead of
QFileDialogs for adding files/folders to an archive. Now both folders
and files are added to an archive with a single action in Part, instead
of two separate actions (Add File/Add Folder).

AddDialog has a button which opens a new dialog that allows setting
advanced compression settings (currently only compression level). Since
compression options are used both when creating a new archive and adding
files to an existing archive, a new class called
CompressionOptionsWidget was created which is used by both AddDialog and
CreateDialog.

Differential Revision: D2120
GUI: