Detect multi-volume archives and show info in PropertiesDialog
Two variables were added to ReadOnlyArchiveInterface: A boolean
describing whether the archive is multi-volume and an int used for
storing number of volumes. Two corresponding Q_PROPERTY's were added to
Archive which fetch the info from ReadOnlyArchiveInterface.

The detection of multi-volume archives was fixed in cli7z.

The information is displayed in PropertiesDialog and will be used when
implementing support for creating multi-volume archives.

GUI: