Disable adding files to existing password-protected non-he archives
We don't know the password after opening a non-header-encrypted
password-protected archive. If adding files to such an archive the newly
added files will be unencrypted with no warning to the user and the
resulting archive will be a mixture of encrypted/unencrypted files.

This commit disables adding files to such archives and gives an
explanation in the tooltip for the add files action.

Differential Revision: D2196