Add support for AppImage
AppImages are both executables and ISO images and can be handled by
libarchive. Support for opening AppImages was enabled in the
libarchiveplugin. The mimetype for AppImage has been committed to
shared-mime-info, but has not yet been released. We therefore install a
custom mimetype for AppImage.

Test cases for this mimetype were added to MimeTypeTest and ArchiveTest.

FEATURE: 363209
FIXED-IN: 16.08.0
Differential Revision: D2226