Add multi-volume support
Support for creating multi-volume rar, 7z and zip archives was
implemented. A QDoubleSpinBox was added to CompressionOptionsWidget
which allows setting the volume size in megabytes between 0.1 to 1000.
The size in megabytes is converted by CompressionOptionsWidget to
kilobytes because 7z doesn't support volume sizes with decimals.

Creating a multi-volume archive changes the archive name
(name.part1.rar, name.7z.001, name.zip.001) so we need to re-open the
archive (the first volume) after adding files.

We only support adding files once, so add/delete actions are disabled in
Part if archive is multi-volume and non-empty.

FEATURE: 124180
FIXED-IN: 16.08.0
Differential Revision: D2194
GUI: