Fix preview of nested archives
The assumption that a PreviewJob can remove its temp dir when it's done (commit 41b2512)
was wrong, at least with nested archives.
If we remove the extracted file too soon, there is no way to preview a nested archive or,
worse, a "double-nested" archive.

This patch just delays the cleanup to the Part destructor.

Differential Revision: D2265