Fix endless loop of ListJobs
There is a race condition when ListJob emits the result signal.
Both ArchiveModel::slotLoadingFinished() and Archive::onListFinished() are connected
to this signal. But the former triggers a chain of calls that ends up calling
properties of Archive that are set by the concurrent Archive::onListFinished().
An example is Archive::encryptionType() called by Part::updateActions().

This results in an endless chain of ListJob instances.

This patch is only a temporary workaround for the upcoming 16.08 stable
release, since there is no time to test and merge a proper fix (see T3300)

Differential Revision: D2272
Task: T3296