Fix handling of drag'n'drop adding files to archives
Multi-volume archives are set to read-only (only when non-empty) in
Archive class. This is needed because rar and 7z don't support modifying
such archives.
Also, a KMessageWidget is now shown when user drags files to a read-only
archive or an encrypted archive when password is unknown.

Unfortunately, we cant set encrypted archives with unknown password to
read-only, because we still want to support deleting files from them.

Differential Revision: D2274