Add testcases for multivolume rar and 7z archives to ArchiveTest
We test two additional properties in testProperties: isMultiVolume and
numberOfVolumes. These properties are only correct after listing an
archive so we now call listIfNotListed() in isMultiVolume(). This change
required isMultiVolume() and isReadOnly() to be made non-const.

The testcases revealed a bug in Archive::completeBaseName(), where the
basename was not generated correctly for multi-volume archives. This is
also fixed now.

Differential Revision: D2291