Drop unused event loop code
The event loop in the QThread's run() method was only necessary when
CliInterface was running in another thread, but this is no longer
happening. This event loop does not even start because when doWork()
returns, `result` has already been emitted, resulting in q->isRunning()
being false.

This also means that we can drop m_isRunning, removing a possible
source of race conditions.