Add support for application/vnd.rar, fixes opening RAR with shared-mime-info 1.7
shared-mime-info 1.7 introduced the official IANA mimetype
`application/vnd.rar` and degraded `application/x-rar` to alias.

For now we need to register both mimetypes. Once distributions start
shipping shared-mime-info 1.7, we can drop the alias.

For this reason, we cannot have a reliable test case in mimetypetest.
Also, in `PluginLoader::sortByComment()` we need to check whether
mimetypes are valid, because `application/vnd.rar` might not be in
`QMimeDatabase()`.

BUG: 368786
FIXED-IN: 16.08.2

Closes T3729