Fix drag-and-drop extraction of parentless files
In some archives parent folders of some files might not be part of the
archive themselves, i.e. the archive does not contain separate entries
for them. In this case ArchiveModel creates ArchiveDirNodes for the
folders. However, the FileName and InternalID properties of these
folders do not get a trailing slash. This results in faulty root nodes
leading to e.g. broken drag-n-drop extraction of files within these
folders.

This diff simply appends a traling slash to the properties in
ArchiveModel::parentFor().

BUG: 367549
FIXED-IN: 16.08.2
Differential Revision: D2909