Register jobs in KIO's jobtracker
This makes all of Ark's job types show up in system tray of Plasma. The
progress of the jobs is then visible in system tray and they can also be
stopped from system tray. Additionally, the user gets a notification
when jobs complete.

Ellipsis were removed from description of LoadJob ("Loading archive...")
to be consistent with description of the other jobs.