Process entries while listing archives
When listing archives Ark currently waits until the LoadJob is done to
process the entries. This causes a lag (>7 seconds for large archives on
fast systems) after the LoadJob has completed until the ArchiveView is
ready.

This diff causes ArchiveModel to start processing entries while the
LoadJob is running. This results in the view being ready <1 second after
LoadJob completes.

Differential Revision: D2948