Fix KMessageWidget not disappearing after adding files
When creating a new archive a KMessageWidget is shown with the text:
"The archive <name> will be created as soon as you add a file."
This widget should be hidden as soon as the user adds a file.

This is already fixed in Applications/16.08 branch and also was
previously in master (commit 50f0cbb42dd2), but somehow got reverted
in the merge in commit 7eb3304db549.