clirar: Detect error when unrar version is too old
Unrar 3 and 4 cannot open RAR5 archives. Currently, when opening a RAR5
archive with these old versions there is an uncaught error message which
ends up in the comment widget.

This commit adds a check for these error messages in clirar and emits an
error signal with a proper message.

CliInterface was changed so that handleLine() passes on the boolean
return value of readListLine() to readStdOut(). This way, we can stop
readStdOut() from handling further lines when readListLine() returns
false. This was needed to avoid multiple error signals being emitted.
Previously, the return value of readListLine() was unused.

Unittests were added for both unrar 3 and 4.

Differential Revision: D2993