Detect compression method
A new property was added to Archive to store detected compression
method(s). Currently, it's only displayed in PropertiesDialog, but the
plan is to also use it to select a sane default compression method when
adding files to an existing archive. Code was added to all plugins for
detecting the compression method(s) when opening an archive and a signal
added to ReadOnlyArchiveInterface to set the property.

Many archive types support multiple compression methods in same archive,
so the property is a QStringList.

Differential Revision: D2987