Turn extraction/compression options into classes
Currently CompressionOptions and ExtractionOprions are both QHash typedefs,
which means they are the same thing for the compiler.
Currently we even pass CompressionOptions objects where ExtractionOptions objects are expected.
Both types are changed into proper classes, so that the compiler can
detect this class of bugs.

While at it:

- The default value for the `PreservePaths` option was false, now is true because it's
  the more common case.
- The `RemoveRootNode` option was redundant, it was only used together with the
  `DragAndDrop` one.
- The `FollowExtractionDialogSettings` was only set but never read, so we can drop it.

Differential Revision: D3039
Task: T2137