movetest: fix root directory test cases
We are not touching the top-level dir1/ and dir2/ folders, so we need to
remove them from the list of target entries. In the second test case we
also need to fix the destination entry, because we are moving only one
folder and in this case we need to include its name in the destination.

GIT_SILENT