Allow extraction of read-only archives
Currently we use `KFileItemListProperties::supportsWriting()` to check whether a folder
is read-write and can be used as extract destination, but this function returns
false if the archive is read-only. A read-only archive can still be extracted,
so we need to implement our own check for the writability of parent folders.

BUG: 370948
FIXED-IN: 16.08.3

Differential Revision: D3074