Drop TestHelper test functions
copytest and addtest rely on TestHelper functions, which are untested
and thus not reliable for test purposes. Instead we manually define the
expected paths of the entries created by the jobs.
A new base class is added, since movetest/copytest/addtest share the
code that sets up the test rows.

Closes T4073