Filter out duplicate plugins
This prevents problems when Ark is installed both from package manager
and in a custom prefix.

Plugins are loaded on a first-come first-served base. This means that if
the QT_PLUGIN_PATH env variable is defined, only the plugins in the
first path defined in that variable will be loaded.