Fix crash when aborting LoadJobs
LoadJob::onNewEntry() receives an Entry pointer. When we kill a LoadJob, the
libarchive plugin deletes all the emitted entries, so onNewEntry() (which is
queued in the events) gets a pointer that points to already free'ed memory.

This patch just delays the deletion of entries by using deleteLater().
This ensures that all the pending slots are executed first.

Differential Revision: D3143