Fix entry counter in libarchiveplugin when overwriting entries
When adding files to an archive with libarchiveplugin and the
destination files already exist (so they get overwritten), the entry
signal is still emitted and hence m_numberOfEntries gets wrongfully
incremented. This diff decrements the counter when an entry already
exists.

Overwriting of entries is disabled for CopyJob and MoveJob, so there is
no problem in those cases.

Differential Revision: D3273