Implement support for setting encryption method
Needed infrastructure changes were done in ArchiveFormat, CliProperties
and other classes. A combobox was added to CreateDialog which enables
setting the encryption method.

Currently, only the Zip format when using cli7z supports multiple
encryption methods. For other formats that support only one encryption
method, the method is shown in the combobox.

This diff also enables detection of encryption method when opening
archives which is then shown in PropertiesDialog. To this end a new
signal was added to ReadOnlyArchiveInterface (encryptionMethodFound).
The similar compressionMethodFound signal was modified to only take one
string instead of QStringList.

Differential Revision: D3303
GUI: