mainwindow: properly disable 'Open Recent' action
KRecentFilesAction::addUrl() makes the action enabled. This is called in
MainWindow::openUrl() after `busy()` has been emitted by the Part, which
means the action is enabled while a LoadJob is running.

If we use a queued connection instead, addUrl() will be called first and then
MaindWindow::updateActions() will disable the action as expected.

BUG: 365200
FIXED-IN: 16.12.0