Improve error handling for CliPlugins
This commit removes the ExtractionFailedPatterns regexp in
CliProperties. Instead we take care of the error handling in the
individual plugins by adding a pure virtual readExtractLine() method to
CliInterface (similar to readListLine()). This way we can provide custom
error messages to Part, instead of the generic "Extraction failed due to
an unexpected error" (this message was also used for ListJob).

Differential Revision: D3142