Don't show two progress bars with batch extractions
BatchExtractJob runs a LoadJob first and an ExtractJob after, which results in two
different progress bars in the notification tray. This patch changes how the
percentage is computed: the first 50% is from the LoadJob, the 2nd 50% from the
ExtractJob. This ensures that only one progress bar shows up while the wrapper job
is running. This will only work if the interface is able to report progress for
both LoadJobs and ExtractJobs (currently only libarchive and clirar).

Differential Revision: D3518