Fix resizing of infopanel
KSqueezedTextLabel doesn't rich text. This causes the used <b> and </b>
tags to sometimes become visible when the infopanel is reduced in size.

This is fixed by splitting the text into two parts: The left-most is a
QLabel (because it has predictable horizontal size) which holds the
title and the right-most is a KSqueezedTextLLabel and holds the value.
The left-side one can then be set to bold font.

Differential Revision: D3523