pluginmanager: cache list of preferred plugins
PluginManager::preferredPluginsFor() is an hot path as it can be called
thousands of times from the extractfileitemaction plugin.
We can save the list of preferred plugins in a cache, instead of
re-computing it over and over again.

Tested in Dolphin by right-clicking:
* 1K files: no delay
* 10K files: ~1 sec. of delay
* 50K files: ~4 sec. of delay

The (empty) test files were created with:
    printf '%s ' {1..1000} | xargs touch

BUG: 372999
FIXED-IN: 16.11.90

Differential Revision: D3527