Implement search function
An action was added to Archive menu and is also shown in the default
toolbar. The search is mediated through KRecursiveFilterProxyModel
instead of QSortFilterProxyModel because the latter does not recurse
nested models. This adds a new dependency on the KItemModels framework.

The search is always case-insensitive.

The search bar is displayed above the QTreeView and contains a close
button. An eventfilter was installed on Part to catch the escape
keypress to close the search bar.

FEATURE: 188197
FIXED-IN: 17.04.0
Differential Revision: D3573
GUI: