Introduce settings page to configure plugins
This commit adds a page in the settings dialog where the user can enable/disable
plugins. The page contains a QTreeWidget which is not supported by
KConfigDialog out of the box, so some glue-code is necessary to
properly handle the Apply/Default buttons. ArkSettings stores a list of
disabled plugins (by default, all plugins are enabled).
If a plugin is not valid because of missing executables, the respective
item in the view is disabled and a warning is displayed in a tooltip.

Differential Revision: D3716
Task: T1984
GUI: