Add isSingleFile property
An archive is single-file if contains only one file. The Archive class
only holds its numberOfEntries() but it doesn't know if they are files
or folders. However, if there is only one entry and isSingleFolder() is
false, that entry must be a file.
This is ensured by the expanded loadtest.