Fix archivemodel columns with singlefile archives
ArchiveModel was setting 0 as Size with singlefile archives,
because the singlefile plugin doesn't report the uncompressed size.

This patch removes the Size column if the size property is 0 but the
compressedSize one is higher than 0, which cannot happen.
As compressedSize for singlefile entries, we just set the size of the
archive.

BUG: 376225
FIXED-IN: 16.12.3

Differential Revision: D4536