Fix a I18n problem caused by using i18n and i18np in two places.
Summary:
There are two entries with "Extracting one file". The one is called with
i18n and the other is called with i18np. Thus it confuses the translation
to use "Extracting %1 files" in i18n() version which is a problem for
languages with "Plural-Forms: nplurals=1; plural=0;". The actual result
before this fix will be using the "%1" without arguments.

Test Plan: Manually tested.

Reviewers: rthomsen, elvisangelaccio

Reviewed By: elvisangelaccio

Subscribers: kde-utils-devel, #ark

Tags: #ark

Differential Revision: https://phabricator.kde.org/D5511