Fix context menu position on Wayland
On Wayland popup menus without a parent won't have a transient parent
window, which is required to show the menu in the right position.

The context menu in the archiveview is created by kxmlgui, without a
parent because it was a top-level <Menu> tag. If we define it as child
of the <MainWindow> tag, we ensure that the main window will be used as
parent of the menu.