Make ArkViewer a KPart mainwindow
This allows us to call createGUI() on the KPart we are embedding. As a
side effect, we are now able to show the popup menu of the ArkPart,
which is a stand-alone popup menu created by kxmlgui, when previewing a
nested archive in the ArkViewer. Same for the menubar and toolbar of the
embedded parts.

By using a KMainWindow (instead of a QDialog) we also gain some
features, for example we get size restoring for free (thanks to
setAutoSaveSettings()), without the need of manually using
KWindowConfig.

BUG: 380439
FIXED-IN: 17.07.80

Reviewers: #ark, fvogt

Differential Revision: D6075