libzip: preserve permissions when extracting files
Summary:
The zip format reserves the 2 leftmost bytes of the extended file
attribute field (32 bits) for the UNIX permissions.
If the zip archive was crated with such metadata, we can just use the
libzip api to retrieve the extended file attribute and extract the
mode_t value from it.

BUG: 381499
FIXED-IN: 17.07.80

Test Plan: This fixes the 2 failing libzip test cases in ExtractTest::testPreservePermissions().

Reviewers: #ark, rthomsen

Subscribers: kde-utils-devel

Tags: #ark

Differential Revision: https://phabricator.kde.org/D6408