Don't add invalid URLs to the recent files
Part::openFile() used to return false if the URL of the part was not a
valid archive. This is no longer the case as nowadays we run an async
LoadJob. So MainWindow cannot use the return value of openFile() to
choose whether to add the URL to the recent files. Instead, we connect
to the part's completed() signal which is emitted only if the LoadJob
ends without errors.