part: update window title whenever the URL changes
This fixes a small regression in commit 20c09f3c43.
A ReadOnlyPart emits `completed()` if `openUrl()` returns true, so we
lost a call to `setFileNameFromArchive()`.
We fix that by connecting this slot to the `urlChanged()` signal.

Tested by opening some archive and then an header-encrypted archive, and
making sure the window title had the name of the second archive while the
password dialog was open.