libzip: Improve test integrity function
The old test implementation only opens the archive with some additional
checks, but fails to detect archives with CRC errors for entries.

We now read each entry and check the CRC checksum against the one given
in the file header. We use zlib's crc32() function to compute the
checksum.

CCBUG: 384622
Differential Revision: D8400