Fix threading issue
`CreateJob::doWork()` might be executed in another thread (if we are
using the libarchive or libzip plugin), so any QObject created there
cannot use `this` as parent (since the CreateJob had been created
in the main thread).

We don't actually need to create an Entry object here, since all the
interfaces have checks for null destination entries.

Fixes T7206