Stop crashing when all read-write plugins are disabled
CreateDialog assumes that `m_supportedMimeTypes` is never empty, it will crash otherwise.

If that's the case (i.e. when all the read-write plugins have been disabled),
we just disable the Compress menu item and the New action (as the
CreateDialog would just be useless).

For the New action, we also need to check whether the settings have been
changed and update its enabled status accordingly.

BUG: 390690
FIXED-IN: 17.12.3