clizip: fix list after adding files to encrypted archive
zipinfo doesn't support the -P switch (since zip doesn't support
header-encryption), so running zipinfo with e.g. -P1234 fails.

This is a problem when we list the archive after adding new files, if the
archive has been created as encrypted. Due to this problem, the file is
actually added but the view doesn't show it.

This patch substitutes the password switch only if the plugin supports
header-encryption (regardless the actual encryption type of the
archive).