libzipplugin: Simplify registration of progress callbacks
Summary:
Use zip_register_progress_callback_with_state() that was
introduced in libzip 1.3 (we already require this version).

Pass 0.001 as precision because this value is used for
old zip_register_progress_callback() inside libzip.
Also pass nullptr as a free function, because 'this' doesn't
need to be freed.

Reviewers: #ark, elvisangelaccio

Reviewed By: #ark, elvisangelaccio

Subscribers: elvisangelaccio, kde-utils-devel, #ark

Tags: #ark

Differential Revision: https://phabricator.kde.org/D16276