Add support for Zstd-compressed Tar-archives
Libarchive 3.3.3 was released on 2018-09-04 and added support for the
Zstandard compression filter. This compression method is designed to
give similar compression ratio as gzip but with much greater
(de)compression speeds. Libarchive uses the libzstd.so library, but
seems to use the zstd binary as fallback if it wasn't built with
libzstd.

This diff enables the filter in the libarchive plugin. There is still no
mimetype in shared-mime-info, so a custom mimetype for zstd-compressed
tar archives is added. This necessitated re-adding functionality in Ark
for custom mime-types (was removed in commit
73cd5e8d2e7b347d84438b23afdc892eddb78ed2).

Support is only enabled when built with libarchive>=3.3.3.

FEATURE: 384040
FIXED-IN: 18.12.0
Differential Revision: D15369