Stop defining empty DEPENDENCY_TOOL_ARGS
On !APPLE systems we were defining an empty DEPENDENCY_TOOL_ARGS macro,
resulting in a clazy warning due to `QStringLiteral(DEPENDENCY_TOOL_ARGS)`.

Now we define this macro only when needed, and we check with an #ifdef if we
need to use it. As a bonus point, we avoid to pass to QProcess::setArguments()
a list whose first element is empty.