Remember part URL to delete the tempfile even if the user used save as.
Summary:
Using the part URL at closing time is really dangerous: if the user used
"save as" (e.g. for an embedded PDF), we'll be deleting the file that
the user just saved!!

BUG: 402529

Test Plan:
launch ARK and open a ZIP file containing a PDF, click on the PDF,
an embedded viewer appears. From there, use "Save As..." to save the PDF
somewhere on your local disk. Close the embedded viewer, close ark.
It doesn't delete the saved file anymore.

Reviewers: elvisangelaccio, broulik

Reviewed By: elvisangelaccio

Subscribers: kde-utils-devel, #ark

Tags: #ark

Differential Revision: https://phabricator.kde.org/D18256