Improve performance of cliplugins
The CliPlugins are really slow when listing and extracting archives.
This is due to several QRegularExpressions being matched on each line of
output.

This diff removes all of the QRegularExpression matchings and uses
string comparison instead (e.g. startsWith(), contains() and ==()). This
necessitated moving some checks from the CliProperties class to the
individual CliPlugins, because each plugin has different requirements
for which string comparison methods should be used.

Differential Revision: D19766