Fix progress info for CreateJob
Progress info is currently broken when creating archives using
CreateJob. This happens e.g. when the user creates an archive
through the KFileItemAction in Dolphin.
This is caused by CreateJob using a nested AddJob and the percent
signal of the AddJob was not forwarded to CreateJob.

It was encessary to use QOverload since KJob has both a signal and a
member function called percent.

BUG: 382599
FIXED-IN: 19.04.0
Differential Revision: D19953