Handle locked RAR archives
RAR archives can have a "locked" property essentially making them
readonly. Currently, Ark doesn't handle this properly and fails silently
if the user tries to modify the archive (e.g. adding files).

This diff adds a virtual method ReadOnlyArchiveInterface::isLocked()
with a default implementation returning false. The method is overridden
by the RAR plugin which checks if the archive is locked when listing.
ReadWriteArchiveInterface::isReadOnly() checks isLocked() when
determining the readonly status of the archive.

BUG: 368851
FIXED-IN: 19.04.0
Differential Revision: D20122