Do not use Q_DECLARE_METATYPE(KPluginMetaData) with KF 5.60+
It is already defined in kcoreaddons since 318b0e75242931ea420caf3e7123d54eb8351c51
and will break compilation.

Use a version macro instead, so that it is correctly used only with KF
5.59 or lower.

I put this in stable branch in case people will use 19.04 with 5.60
(likely), but feel free to revert in case it is not appropriate.