libarchive backend: display permissions in octal format
Summary: Currently when libarchive backend is used to read archive, permissions are displayed in decimal format. But a lot of userspace tools show and use permissions either in text format or in octal number format for convenience, not in a decimal number format.

Test Plan:
Open an archive which ark uses libarchive for opening. For example, https://github.com/KDE/ark/archive/v19.12.3.tar.gz

For directories permissions should be displayed as '40775' instead of '16893'.
For non-executable files permissions should be displayed as '100664' instead of '33204'.

Reviewers: #ark, elvisangelaccio

Reviewed By: #ark, elvisangelaccio

Subscribers: kde-utils-devel

Tags: #ark

Differential Revision: https://phabricator.kde.org/D28007