Fix extraction to directories with unicode names when using the Extract file menu
Before this patch, BatchExtract::setDirectory() would ignore directories whose path
contains a tab or an emoji, because QUrl::toDisplayString() returns a percent-encoded name
(regardless of FormattingOption) which does not get percent-decoded.
(see https://stackoverflow.com/questions/24343582/convert-qurl-with-percent-encoding-into-string/24394362#24394362)

We now use QUrl::toLocalFile() instead, which does not percent-encode the name, and we warn the user
if the destination is not local (which is not supported by Ark).

This bug was only present on the extract file menu (via dolphin context menu).

BUG: 408186
FIXED-IN: 20.11.80