Fix handling of MIME subtypes of compressed tar archives.
Change determineMimeType in mimetypes.cpp to use
QMimeType.inherits() instead of QMimeType.operator== when
checking if the MIME type determined from the file extension
is that of a compressed tar archive.
This fixes the handling of files that use MIME subtypes of
application/x-compressed-tar etc. This is needed to be able to
extract ALPM packages used by Arch Linux.
See bug 416092.