Show a "Welcome screen" when launched without an archive
This merge request replaces the window contents when the Ark application
is started without giving it archive to open. The current user interface
has all its widgets (except the list view) in disabled state and the only
way to open or create archive is to use menu bar (not shown in pictures)
or keyboard shortcuts. It can be  hard to figure out what to do for
inexperienced users.

The new user interface adds two buttons directly to the center of window:
one for opening archive and the second for archive creation. When archive
is opened, this screen is replaced by the normal user interface.

This changes only Ark's main window, not Ark KPart, so it will not break
any existing usages of the KPart.