app/extractfileitemaction: Ask for password again if incorrect when extracting
Currently only when extracting zip files, Ark will ask for the password
again if the entered password is incorrect. This is partially due to the
extracting process directly quitting instead of asking the user for a new
password.

Ark should be able to handle that case and ask for a new password again
if the entered password is wrong.

If `ShowDialog` option is set, `AllowRetryPassword` is not set. Because if
`AllowRetryPassword` is set, the dialog will be displayed every
time before the password dialog is displayed.

BUG: 392555