Do not highlight file after compression
Dolphin opens a new tab whenever the file is compressed and
is not smart enough to highlight it in the currently open view.
Because people are annoyed by it and the alternatives are complex to implement,
removing it is the best solution.

BUG: 440663


(cherry picked from commit 28f2ef4b22f53200cb8789dbc8fe8ecdba3a377f)