Group identically-scoped buttons on Toolbar
Right now Ark's viewer window toolbar contains buttons with different
scopes: some operate on the full contents of the view, while others
operate on the set of selected items. The ones with a selection-scope
correctly disable themselves when nothing is selected, which is good.
However they are not grouped together; instead they are intermixed with
the global-scope actions. This is slightly confusing.

This commit fixes that by grouping the global-scope actions and the
selection-scope actions together, and separates the groups with a
separator.