Clarify selection-scope Open actions
Ark currently has two "Open" actions: one is global-scope (open a new
archive file) while the other is selection-scope (open the selected file
in an external program). They both have identical text and icons, and
are disambiguated only by their tooltips and locations in the menu
structure. However this does not help when the action is added to the
toolbar, as it is by default. In this circumstance, its purpose is very
unclear as it has the same name and icon as a typical "Open a new file"
action, but it doesn't open a new file.

To clarify the situation, this commit makes the following changes to the
selection-scope open actions:
1. Use longer text that explains what the action will do, and delete the
   special tooltip text
2. Use the document-export icon, which uses the more appropriate visual
   metaphor of "get the document out of this app".